requirements.txt
streamlit
pandas
plotly
Pillow
gspread
gspread_dataframe
oauth2client
openpyxl
